Rename files with camel case format, fix inline comments to top (separate) line comments.
