+++ index.js
@@ -1,7 +1,7 @@
 /*!
  * Copyright (c) 2025 Digital Bazaar, Inc. All rights reserved.
  */
-export {CameraScanner} from './lib/camera-scanner.js';
-export {OpticalScanner} from './lib/optical-scanner.js';
+export {CameraScanner} from './lib/cameraScanner.js';
+export {OpticalScanner} from './lib/opticalScanner.js';
 export * from './lib/plugins/index.js';
 export * as cameraUtils from './lib/utils/camera.js';
