+++ lib/cameraScanner.js
@@ -12,7 +12,7 @@ import {
 } from './plugins/index.js';
 import {EventEmitter} from 'events';
 import {getDynamsoftLicense} from './config.js';
-import {OpticalScanner} from './optical-scanner.js';
+import {OpticalScanner} from './opticalScanner.js';
 
 const logger = createLogger('[CameraScanner]');
 
@@ -27,11 +27,14 @@ export class CameraScanner extends EventEmitter {
 
     // Extract configuration options
     const {
-      scanType = options.scanType || 'barcode', // 'mrz' | 'barcode' | 'auto'
+      // scanType = 'mrz' | 'barcode' | 'auto'
+      scanType = options.scanType || 'barcode',
       mrzMode = options.mrzMode || (scanType === 'mrz' ? 'camera' : 'element'),
       licenseKey = getDynamsoftLicense(),
-      scanMode = options.scanMode || 'first', // 'first' | 'all' | 'exhaustive'
-      formats = null, // Explicit format override
+      // scanMode = 'first' | 'all' | 'exhaustive'
+      scanMode = options.scanMode || 'first',
+      // Explicit format override
+      formats = null,
     } = options;
 
     // Validate scanType
@@ -71,7 +74,8 @@ export class CameraScanner extends EventEmitter {
     // ===== INTERNAL STATE =====
     this._stream = null;
     this._videoElement = null;
-    this._container = null; // Store container for plugin options
+    // Store container for plugin options
+    this._container = null;
     this._opticalScanner = null;
     this._isScanning = false;
     this._torchState = false;
@@ -119,7 +123,8 @@ export class CameraScanner extends EventEmitter {
   _getFormats(scanType) {
     // Check for explicit format override first
     if(this.config.formats !== null) {
-      return this.config.formats; // returns ['pdf417_enhanced']
+      // returns ['pdf417_enhanced']
+      return this.config.formats;
     }
 
     // Otherwise, use scanType defaults
@@ -172,7 +177,8 @@ export class CameraScanner extends EventEmitter {
           pluginOptions.pdf417_enhanced = {
             licenseKey,
             useDynamsoft: true,
-            parseDL: true // Parse driver license data
+            // Parse driver license data
+            parseDL: true
           };
           break;
 
@@ -225,7 +231,8 @@ export class CameraScanner extends EventEmitter {
           licenseKey,
           mrzMode,
           scannerConfig: {
-            container, // Critical for Dynamsoft native UI
+            // Critical for Dynamsoft native UI
+            container,
             scannerViewConfig,
             resultViewConfig
           }
@@ -380,7 +387,8 @@ export class CameraScanner extends EventEmitter {
         videoReady: !!this._videoElement,
         scanType: this.config.scanType,
         mrzMode: this.config.mrzMode,
-        autoScanStarted: false // Don't auto-start if already running
+        // Don't auto-start if already running
+        autoScanStarted: false
       };
     }
 
@@ -425,7 +433,8 @@ export class CameraScanner extends EventEmitter {
             console.error('Auto-scan completed or failed:', error.message);
             // let the scan method handle errors through events
           });
-        }, 100); // Small delay to ensure start() completes first
+        // Small delay to ensure start() completes first
+        }, 100);
 
         autoScanStarted = true;
       }
@@ -437,7 +446,8 @@ export class CameraScanner extends EventEmitter {
         scanType: this.config.scanType,
         mrzMode: this.config.mrzMode,
         formats,
-        autoScanStarted // indicates if auto-scan was initiated
+        // indicates if auto-scan was initiated
+        autoScanStarted
       };
 
       return result;
@@ -700,7 +710,8 @@ export class CameraScanner extends EventEmitter {
       // avoid creating options in between.
 
       const results = await this._opticalScanner.scanAny(scanSource, {
-        mode: 'first', // First successful format wins
+        // First successful format wins
+        mode: 'first',
         pluginOptions: this._pluginOptions,
         timeoutMs,
         signal
@@ -1217,7 +1228,8 @@ export class CameraScanner extends EventEmitter {
       try {
         isDebugEnabled() && console.log(
           ...logger.prefix('Attempting to recover original camera...'));
-        await this.start(this._container); // Restart with default camera
+        // Restart with default camera
+        await this.start(this._container);
       } catch(recoveryError) {
         console.error('Failed to recover camera:', recoveryError);
       }
@@ -1265,8 +1277,10 @@ export class CameraScanner extends EventEmitter {
       const filePluginOptions = this._buildPluginOptions(
         formats,
         this.config.licenseKey,
-        null, // No container needed for file scanning
-        'file' // Force file mode for MRZ
+        // No container needed for file scanning
+        null,
+        // Force file mode for MRZ
+        'file'
       );
 
       // Scan first file (can be extended to scan all files)
