+++ lib/opticalScanner.js
@@ -507,13 +507,15 @@ export class OpticalScanner {
   */
   async scanAny(source, options = {}) {
     const {
-      timeoutMs = 0, // Extract timeout from options
+      // Extract timeout from options
+      timeoutMs = 0,
       ...restOptions
     } = options;
     return this.scan(source, {
       ...restOptions,
       formats: this.getSupportedFormats(),
-      mode: 'first', // Try formats until one succeeds with data
+      // Try formats until one succeeds with data
+      mode: 'first',
       timeoutMs
     });
   }
@@ -614,7 +616,8 @@ export class OpticalScanner {
   *  occurs.
   */
   async _waitForAll(promises, signal) {
-    signal?.throwIfAborted(); // Immediate exit if aborted
+    // Immediate exit if aborted
+    signal?.throwIfAborted();
 
     try {
       const results = await Promise.allSettled(promises);
